[SPARK-39964][SQL] DS V2 pushdown should unify the translate API
